+++ tests/test_batch.py
@@ -1,12 +1,15 @@
 from uuid import uuid4
+from unittest.mock import patch
 import pytest
 
 from sdk import SDK, DeviceType
 from sdk.device.configuration import BaseConfig, EmuFreeConfig, EmuTNConfig
+from sdk.authentication import FakeAuth0GoodAuthentication
 
 
 class TestBatch:
     @pytest.fixture(autouse=True)
+    @patch("sdk.client.Auth0TokenProvider", FakeAuth0GoodAuthentication)
     def init_sdk(self, start_mock_request):
         self.sdk = SDK(
             username="me@test.com", password="password", group_id=str(uuid4())